backend/pyproject.toml
[tool.poetry]
name = "avatar-mongodb"
version = "0.1.0"
description = "Avatar - MongoDB Hackathon"
authors = ["ecohen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic= "1.11.3"
annotated-types= "0.5.0"
anyio= "3.7.1"
bcrypt= "4.0.1"
click= "8.1.7"
colorama= "0.4.6"
ecdsa= "0.18.0"
fastapi= "0.103.0"
FastAPI-SQLAlchemy= "0.2.1"
greenlet= "3.0.3"
h11= "0.14.0"
idna= "3.4"
Mako= "1.2.4"
MarkupSafe= "2.1.3"
passlib= "1.7.4"
pydantic= "2.5.2"
pydantic-extra-types= "2.4.0"
pydantic-settings= "2.0.3"
pydantic_core= "2.14.5"
PyMySQL= "1.1.0"
python-dotenv= "1.0.0"
python-jose= "3.3.0"
rsa= "4.9"
six= "1.16.0"
sniffio= "1.3.0"
SQLAlchemy= "2.0.20"
starlette= "0.27.0"
typing_extensions= "4.7.1"
uvicorn= "0.23.2"
inflect = "^7.2.0"
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.9.0"
pytest = "^8.1.1"
requests = "^2.31.0"
bs4 = "^0.0.2"
llama-index = "^0.10.27"
pymongo = "^4.6.3"
llama-index-vector-stores-mongodb = "^0.1.4"
llama-index-embeddings-openai = "^0.1.7"


[tool.poetry.scripts]
dev = "main:dev"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
